vehicle value format
