+++ policyengine_us/parameters/gov/states/ca/cdss/capi/resources/vehicle/disregard.yaml
@@ -1,8 +1,8 @@
-description: California counts the value of a car over this threshold as resources for the Cash Assistance Program for Immigrants.
+description: California counts the value of a vehicle exceeding this threshold as resources for the Cash Assistance Program for Immigrants.
 metadata:
   unit: currency-USD
   period: month
-  label: California CAPI resource limit car value threshold
+  label: California CAPI vehicle value disregard
   reference:
     - title: California Eligibility and Assistance Standards Chapter 49-040 (.3)(c)(5)
       href: https://www.cdss.ca.gov/Portals/9/CAPI/CAPI_Regulations-Accessible.pdf#page=38
