+++ policyengine_us/tests/policy/baseline/gov/states/ca/cdss/capi/resources/ca_capi_countable_vehicle_value.yaml
@@ -1,0 +1,15 @@
+- name: Vehicle below resource threshold
+  period: 2024
+  input:
+    household_vehicles_value: 4_500
+    state_code: CA
+  output:
+    ca_capi_countable_vehicle_value: 0
+
+- name: Vehicle above resource threshold
+  period: 2024
+  input:
+    household_vehicles_value: 6_000
+    state_code: CA
+  output:
+    ca_capi_countable_vehicle_value: 1_500
